refactor: split page and catalog components
